Eliminate unconditional waiting in node join/shutdown
When joining and shutting the node down there are some places where we
unconditionally wait before checking if we should wait any longer.
This has been eliminated, the check is done before the first wait as
well. On node shutdown we always wait for response on pending
invocations but there might not be any invocations in which case we
can just finish the invocation system shutdown without waiting.