Change default flush behavior for WriteFileP
Reason: the flush occurs after draining the Inbox, not after each item. If the
inbox is full, the difference is negligible (benchmark shows, that it's even a
bit faster). For the sporadic output scenario, even though it reduces
performance dramatically (up to 20x), we want to see the item in the file early.