NonBlockingSocketWriter volatile write reduction
In the previous code for every packet written, a volatile write to the pendingFrame is done.
So if there are 10 packets that can be written in a single 'handle' call, it will still lead to
10 volatile writes.

This PR reduces that to a single volatile write. So we only store to the pendingFrame if we
are done filling the outputbuffer.