Temporarily persist joined member uuids
When a member joins and leaves the cluster very quickly, a stale join request can be processed by master and the member is added back to the cluster by mistake.

With this change, we temporarily persist the member uuids that has joined in the 'joinable' cluster state so that stale join requests will be ignored. If the cluster is not in 'joinable' state, previous members can leave the cluster and join back with hot restart.