Fixes memory leak caused by thread group usage
Since ThreadGroup is adding itself to its parent list in constructor,
one has to call ThreadGroup.destroy to remove it back from the list.
ThreadGroup.destroy has to be called after all threads are finished.
We were not waiting for threads to finish before, to avoid memory leak
it seems we have to. But there is a caveat.

It seems impossible to wait for all threads to finish succesfully
because the shutdown thread is also waiting itself to finish.
There is no way to create a Thread adding itself to ThreadGroup
of parent. Hence shutdowns initiated in one of our own threads
has to wait itself to finish.

As a solution, this pr removed the ThreadGroup usage completely.
Justification 1:
We are using ThreadGroups to be able to interrupt all running threads.
As long as we don't create free threads(without executor service), this
interrupting is not necessary. And when we create a thread without executor
service, it is on purpose such as shutdown thread.

Justification 2: A note from Effective Java 2nd Ed., Item 73
"The API that lists the subgroups of a thread group is similarly flawed. While these problems could have been fixed with the addition of new methods, they haven’t, because there is no real need: thread groups are obsolete.
Prior to release 1.5, there was one small piece of functionality that was available only with the ThreadGroup API: the ThreadGroup.uncaughtException method was the only way to gain control when a thread threw an uncaught exception. This functionality is useful, for example, to direct stack traces to an application- specific log. As of release 1.5, however, the same functionality is available with Thread’s setUncaughtExceptionHandler method.
To summarize, thread groups don’t provide much in the way of useful functionality, and much of the functionality they do provide is flawed. Thread groups are best viewed as an unsuccessful experiment, and you should simply ignore their existence. If you design a class that deals with logical groups of threads, you should probably use thread pool executors (Item 68)."

fixes #10394