The SocketChannelWrapper is created earlier for fail fast.
Currently the SocketChannelWrapperFactory is retrieved when it is truly needed, e.g when
a tcp connection is accepted. If there is a configuration error in setting up the
SocketChannelWrapperFactory (e.g. some tls configuration errors) , the exception will
be thrown on the Acceptor thread.

This PR changes that by creating the SocketChannelWrapper early; so when the user
starts a HZ instance, the IO service is also started and now also the SocketChannelWrapperFactory
as well