Prevent leaking ownership info on new arrived node
When a new node arrived it gets ownership info of all clients in
the cluster. It could be the case that a client and its owner member
just left the cluster. In this case, new member gets the ownership info
but it will never be deleted. Because it already missed related events/
operation.

To prevent the leak, owner member is checked against current
memberlist in post join operation. If not in the list, we don't add it
in the first place.