Configurable event sync timeout
It was hardcoded at 5 seconds, but now it is made configurable. This property is needed
to configure the max waiting time for a 'fake synchronous event' to get the outbound
queue of a connection drained. But if the wait is short; the mechanism doesn't provide
sufficient back pressure and you still end up with an overloaded connection.