Adds Channel attribute map
Using the attribute map one can add data to a channel without needing to add fields to that channel.

The protocol-buffer has been removed as explicit field on the channelreaders; and is now stored
as attachement on the channel. The protocol is serverside specific; client doesn't care for
protocol-buffer; so it should not be part of the shared infra-structure.