Speed up Near Cache tests by calling assumeThat() methods first
The assumeThatMethodIsAvailable() method used the NearCacheTestContext
to retrieve a DataStructureAdapter. The test passed the actually used
DataStructureAdapter to the test (dataAdapter or nearCacheAdapter),
because in theory they could be different. In reality they are the same,
so we always had to create the NearCacheContext (which fires up HZ
instances). This slowed down a lot of tests and creates unnecessary
logging.

Since 310 of 784 Near Cache tests are skipped, this should improve the
total test duration significantly.