Improved AwsCloudDiscoveryTest
This test requires the environment variables AWS_ACCESS_KEY_ID and
AWS_SECRET_ACCESS_KEY to be set. This makes it hard to execute the test
locally, since it will fail with a misleading error message.

With this change we check the values upfront by failing if they are not
set on Jenkins and just ignoring the test in other runs. This will make
it easier to get the testsuite running.