Add map and cache event journal
Add event journal for storing and reading mutation events on map and
cache. The events are identified by a long sequence ID.
The events are written in a ringbuffer (event journal) and can be
read from MapProxyImpl, CacheProxy, ClientMapProxy and
ClientCacheProxy. The reading is done by first subscribing and then
calling the read method.
The subscription currently only returns the oldest and newest
sequence ID but it could hold more information in the future, such as
 the subscription ID.