Fixed HazelcastInstanceFactoryTest which left rogue threads alive
The HazelcastInstanceFactoryTest contained a test which throws an
exception in NodeExtension.beforeShutdown(). This prevented a clean
shutdown of the instance, so basically all threads of that instance
were still running. This caused a failure in ThreadLeakTest.

After some discussion we decided that a failure of beforeShutdown() is
not an unsupported scenario (it's just used internally), so we don't
need to fix this with a try-finally-block, but remove the test.

Made HazelcastInstanceFactoryTest to check for thread leaks (just for
all tests together, so we can keep the HazelcastParallelClassRunner).