Removed superfluous initDefaultMaxSizeForOnHeapMaps() for default maps
Since the Near Cache implementations have been unified, we have
different default values for the EvictionConfig.getSize():
* Integer.MAX_VALUE for on-heap IMaps
* 10000 for JCache and off-heap IMaps

We use NearCacheConfigAccessor.initDefaultMaxSizeForOnHeapMaps() to set
the correct default value for IMap and JCache, if no custom value was
set.

The default maps have no NearCacheConfig, so there is nothing to fix.
If a user modifies the default configuration, then the getMapConfig()
call doesn't enter the branch, which creates a new MapConfig and tries
to apply the fix.

The added tests check all combinations of changed and un-changed map
and cache combinations (default + named configurations).