Fail fast when adding non-serializable processor supplier (#463)
Processors have to be serializable. Out of the box processors can become
non-serializable if they are passed a non-serializable lambda. Currently
the job fails if run on multiple instances and the problem is harder to
spot, because it fails when serializing the ExecutionPlan, which might
contain lot of lambda expressions.

This PR makes this problem easier to spot and additionally will fail
even when using single Jet member.