anonymous and inner class support for user code deployment
When classes are loaded via two different class loaders, they can
only access the `public` methods and classes.

Since inner classes are tend to be private, and anonymous classes
can not be public, we made them loaded via  same class loader in
this prd. This way, we will make anonymous and inner classes more
usable when used with user code deployment.

This prd assumes that $ sign is not used as class name for user
classes, although it can be used in class names. It seems that
it is not forbidden, but discouraged.

Related doc http://docs.oracle.com/javase/specs/jls/se7/html/jls-3.html#jls-3.8
`
The "Java letters" include uppercase and lowercase ASCII Latin letters
A-Z (\u0041-\u005a), and a-z (\u0061-\u007a), and, for historical
reasons, the ASCII underscore (_, or \u005f) and dollar sign
($, or \u0024). The $ character should be used only in mechanically
generated source code or, rarely, to access pre-existing names on
legacy systems.
`