Ignore failure when destroying event journal and service is unreachable
This can happen when the ringbuffer service is unreachable because
the node is already shutting down. Unfortunately, we can't get the
reference to the service when creating the event journal
implementation because the implementation is created before the
ringbuffer service. This fix logs an error in case the service is
unreachable and returns.