Fix for class cast exception when running on mule
ClassCastException was there to check whether the hooks listed in META-INF
are actually implements relevant interface or not.
An example to related interface is PortableHook.
Since hook classes are moved in 3.6, now it is hitting this exception and
causing instance to fail fast unncessarily.
To fix the issue we had to gave up on fail fast behaviour. Exception converted
to log to warn the user.

Fixes #10706