Use `find*Config` methods to get ringbuffer config for event journal
The `get*Config` methods add a new configuration based on the "default"
if there is no config with an equivalent name. In a scenario where
maps with different names are requested, the number of configs then
increases and pattern matching on these slows down the event journal
initialisation. Switching to `find*Config` causes the overhead of
config lookup to remain constant.