Fix transactional queue offer/take leak
The transactional queue offer adds an item to the transactional map
but the offer/poll operation expects the item to be in the
(committed) backup map, logs a warning if it is not there and leaves
the uncommitted item as a leak.
This fix checks the transactional map on the backup as well and removes
the reserved item.

Also added some javadoc and removed checked exception on queue
container instantiation.

Fixes : https://github.com/hazelcast/hazelcast/issues/10867

(cherry picked from commit 82f1a5c)