Refactor ringbuffer container creation
The ringbuffer container creation can run into a NPE if the container
is being destroyed simultaneously. This is because of a
check-then-act paradigm : the code checks if there is a container
created and skips config lookup in that case. The problem is that the
container might have been destroyed after the check.
The fix uses a different approach and always calls the ringbuffer
get-or-create method with a non-null argument.
Also renamed some methods to reflect their usage and added javadoc.