Primitive connection qos detection mechanism
Currently we don't have a good rtt measuring system in place. But the
operation heartbeats are send on a fixed interval and not obstructed
by operations.

If there are connection problems between 2 members, the heartbeats will
not be received at the expected interval.

This PRD adds logging of the deviation; if it gets above a certain threshold
it will be printed on warning; or else it will remain in FINE.