Adds a new Network qos diagnostics plugin
This plugin currently works based on the deviation between
operation heartbeats to determine the qos of the network.

Operation heartbeats are processed by their own system and
are priority packets; so no reliance on operation-executor.

If there is a big deviation; then we know there are potential
networking problems. This will hopefully assist with split
brains and otheer complex networking issues.