Eliminate member left handling delay
This is a hardcoded 1111ms delay, which delays triggering invocations
with `MemberLeftException` when a member left is detected. Its purpose
is to wait for an actual response that's possibly sent by left member
and not processed yet. This is just an optimization,
a response may not be sent at all or 1111ms delay may not be
sufficient enough even a response is sent.