Introduce failure detector abstraction and PhiAccuralFailureDetector
Currently Hazelcast has a single failure detector mechanism which is based on a fixed deadline. When a member doesn't any send heartbeat messages until a deadline, it's suspected and marked as failed. Even though this is a simple mechanism, it's not adaptable to environment and it's difficult to define a deadline. If deadline is too big then it takes too much to declare a failing member. If it's too small then it can lead to spurious member removals and cluster splits.

The ϕ Accrual Failure Detector is a more advanced failure detector algorithm which is already used by Cassandra and Akka. It's adaptive to network conditions and doesn't require a predefined deadline. We'll abstract the failure detector mechanism in Hazelcast and provide an implementation of ϕ Accrual Failure Detector.