MapLoader null keys
Fixes #11079

Documentation has been added to the MapLoader.loadAllKeys that no null keys should be returned.

Documentation has been added to MapLoader.loadAll(keys) should not return a map with any null keys.

And a precondition check has been added to the DefaultRecordStore.getRecord to verify that no
null is passed as argument.