Rewrote Near Cache tests to reliably populate the Near Cache
The idea behind this refactoring is to have a race free setup to test
the Near Cache. There is some unreliableness in the event listener
lifecycle and a race between the initial data structure population
(which creates invalidation events) and the Near Cache population.

The easiest solution is to create a member first, populate the backing
data structure and then create the member, lite member or client with
the configured Near Cache. In this case no invalidations are created
from the initial data population, since no event listeners are
registered.