If SSL is enabled, then enterprise should be used.
Currently the SSL configuration is ignored when enterprise isn't loaded. This causes
confusion and potentially can cause security risks.

So instead of ignoring the situation, an exception is thrown (fail fast). This check is
done on both the client and the member.

Fixes enterprise issue: https://github.com/hazelcast/hazelcast-enterprise/issues/1626