This test problably failed due to timing issues.
What happens is that 3 operations are send to a remote box and need to run for 2 seconds. And then
the a check is done if 3 operations are actually running. If there are a million tests running,
then it can happen it takes more than 2 seconds before doing the check (e.g. gc pauses or no
chance to get scheduled).

So the operations are now running for 10 seconds instead of 2.. this shouls reduce the chance of
a spurious failure.