Fix Invocation_TimeoutTest
The problem with this test is that it is time based. Since we are running many tests
in parallel, and the timing are pretty tight; once and a while a timing can fail.

So the simplest thing that can be done now is just double the timeouts and hope the
problem doesn't happen again.