If security enabled and not running enterprise, then error
Currently it is silently ignored which can lead to security issues if by accident
the enterprise jars are missing from the classpath.

It is best to fail fast. So a check has been added on node startup to verify that
enterprise is enabled if security is enabled. Otherwise an exception is thrown
that prevents the node from starting.

fix https://github.com/hazelcast/hazelcast-enterprise/issues/1626