Added create and destroy tests for the Near Cache preloader
Also the ByteBuf used by the NearCachePreloader was changed from a
final field to a normal field. I did some profiling with both variants
and with and without the new tests. The GC profile looks better with
the dynamic field, because the buffer is not propagated to the old gen
so fast. This leads to cheaper GC collections and less total GC time.