Introduce "fail on indeterminate operation state" configuration
If MemberLeftException is received before the ack from the target of a partition, the operation is retried, which could cause the operation to be executed more than once in some racy scenarios. Additionally, if a partition operation has sync backups, the caller waits until acks are received from specified number of backup replicas. If the acks are not received within the configured duration, the operation returns silently.

When this configuration is enabled, the invocation throws IndeterminateOperationStateException if:
1) the operation fails on partition primary replica node with MemberLeftException,
2) there is at least one missing ack from backup replicas for the given timeout duration.

When IndeterminateOperationStateException is thrown, the system does not try to rollback changes. Hence, this new behaviour does not guarantee linearizability.
However, if an invocation completes without IndeterminateOperationStateException when the new configuration is enabled, it is guaranteed that the operation is executed exactly once on the primary partition and specified number of backup partitions.

Invocations which are done with InvokeOnPartitions class are not effected by this new configuration because PartitionIteratingOperation doesn’t wait for sync backup acks under the hood.