Fixes NPE on Nio.close
It can happen that a NioChannel is created but the reader/writer has not been set.
If a close is called, a NPE is thrown since the close if forwarded to the reader/writer.

This PR fixes that by adding a null check on the read/writer before calling close.