Fix scheduled executor leak
The workQueue inside LoggingScheduledExecutor contains wrapped
futures of scheduled tasks. This is because we decorate tasks in
LoggingScheduledExecutor.decorateTask.
When we were cancelling the task, we delegated cancellation to the
delegate which tries to remove itself from the workQueue. The
workQueue though doesn't contain the delegate but rather contains the
decorated delegate. This causes the remove to fail without any
message and causes the leak.
To fix this, after we call cancel() on the delegate, we manually
remove the decorator from the workQueue.

Fixes : https://github.com/hazelcast/hazelcast/issues/11221