Fix ScheduledFutureProxy partition & member listener leak (#11274)
When creating a ScheduledFutureProxy it listens for partition & member loss events.
These listeners only get unregistered when the Future at hand, is disposed. However,
lots of use-cases demonstrate that the original Future is ignored. Instead, calls to
scheduler.getAllScheduled() are utilized to fetch all running tasks, and dispose the
finished ones. This, creates even more proxy objects, all following the same practise,
registering themselves as listeners, therefore they can't be garbage collected.

This patch, introduces a middle-man who acts as the sole & universal listener for those
events, the service itself, and all FutureProxy objects, are weakly referred from within
this class.

The contract of the ScheduledFutureProxy has changed to reflect j.u.c.FutureTask which
uses reference-equality. This also allows the weak reference collection to work as designed.

Note: This change makes the ScheduledExecutorService listener on partition & member loss
events. This is regardless if its in use or not.