Add methods on MapService and CacheService to get object namespace
This is an optimisation to easily get the object namespace from code
paths that either do not have a reference to the MapContainer or
RecordStore or would rather not trigger creating one. The method also
helps in case of refactoring if the type or contents of the namespace
were to change so the change would be localized.