Fix for non smart listeners
When a connection is removed listeners resources are cleaned immediately
on both server and client.
On the client side, when connection is removed,
if clean up is scheduled to happen after new registartions, then we were
falsely removing the event handlers of new registrations.

To fix the order problem between `connection added` and `connection removed`
the task of `connection removed` is moved to `connection added`. Cleaning
up event handlers will be done when a new regitration done for same listener
registration.

getActiveRegistrations method used in tests are also modified to be able
return empty list when connection is removed.

fixes the failure seen in
https://github.com/hazelcast/hazelcast/issues/10235#issuecomment-325973662
related to #10235