Add partition ID to RingbufferWaitNotifyKey
This makes the operation wait/notify key more fine grained.
Previously the key was for all partitions. This meant that for an
read operation for a specific partition to be notified, all other
read operations before it in the wait queue needed to be notified and
processed first.

This causes a form of head-of-line blocking - if a read operation
returns true from shouldWait, other operations in the queue that
might get processed don't get a chance.