Index creation for HD on post-join operation (#11075)
Proper creation of runtime HD indexes on partition-threads after post-join operation. 

Now, the indexes are sent in the map-replication operation for each partition since only this approach guarantees that that there is no race between index migration and data migration. Earlier the index definition used to arrive in the post-join operations, but these operation has no guarantee on order of execution, so it was possible that the post-join operations were executed after some map-replication operations, which meant that the index did not include some data.

Now, each partition thread is responsible for managing the lifecycle of a partitioned-index, thus we can't add an index in the post-join operation directly anymore (also partitioned-index is used in HD only for now, and we are not allowed to touch HD memory on a non-partition thread). That is the reason why we gather all the dynamic post-join index infos for a map in the map-container. Later on they are picked up by the MapReplicationOperation which is spawned for each migrated partition, and the index is added by it for each partition.

MapIndexSynchronizer is an additional anti-entropy mechanism that takes care of the index creation in case the PostJoinMapOperation arrives after some MapReplicationOperations. This is only for the 3.8 to 3.9 upgrade and will be remove in 3.10 codebase, since the race is eliminated by the index definitions being sent in the MapReplicationOperations.