Client heartbeat monitor should only apply for owner connections on non-CPP clients.
Reasoning: the existing condition in client heartbeat monitor was
already checking whether this memberUUID is owner of the given
clientUUID, assuming that there exists a single connection between
a member and a client. However sometimes this is not the case and this
may lead to client heartbeat monitor closing a non-owner connection.