Client fails to reconnect when domain names are used in member list
Make sure that the provided address string is used in the activeConnectionsMap. This fixes the domain name to ip address conversion problem and duplicate connection initiation problem to the same member one with ip address and one with domain name.

We made the assumption that the user will have to use the same host name or ip in the client config if the member is configured with a public ip address.

fixes #11116
fixes #11264
backport of #11226