Log if current cluster quorum value passes quorum value of a jet job
If a job enables split brain protection, JobCoordinationService prints a log when the current quorum value of the cluster becomes bigger than quorum value of the job. The log is printed because the job loses its split brain protection feature in this case. If the cluster splits into partitions, the job can be successfully restarted on both sides.