Fix ConfiguredBehaviourTests
Tests were not waiting for clients to disconnect. It could be the
case that, client did not get disconnections when servers are closed.
In that case tests can fail.

When I add the waits for shutdown/disconnect, I realized that the thread
doing the shutdown (cluster executor thread), waiting itself to shutdown
and timing out with warning. We had an external thread spawn for these
cases before, it  seems that in new development of ClientConnectionStrategy
it is missed. I added that back.

fixes #10810