Skip projection cloning on readFromEventJournal method
The projection was being cloned for defensive reasons. This usually
applies when calling methods using the query engine as user supplied
objects can carry state that will negatively affect the results. In
this case, it is acceptable to clone the user supplied objects since
the method is called only once and all results are returned.
This is not the case for the readFromEventJournal method - it does not
use the query engine and possibly does not suffer from such issues and
the method is called repeatedly to receive new events. This increases
the overhead of cloning.
The fix will skip cloning and emphasises that the caller may clone the
objects once to avoid any unwanted state and to cache the cloned object
to avoid the overhead.

Fixes: https://github.com/hazelcast/hazelcast/issues/11410