Fixed various client class naming violations
An abstract class normally is called this way; and not an impl. Impl is for default
implementations when there is only 1.

Fixes:
ClientInvocationServiceImpl -> AbstractClientInvocationService
ClientListenerServiceImpl -> AbstractClientListenerService

The speciality of a class is normally prefixed in front of the class name:
ClientNonSmartFoobar --> NonSmartClientFoobar
the same goes for the smart.

We use these naming patterns on the server, we should do that as well on
the client to keep naming consistent en predictable.