Catch exception when running task for scanning for metrics
The task is run on a scheduled executor which silently stops the task if
there is an exception. The fix simply logs the exception since we expect
that the exception might only be thrown on the call to
StatisticsAwareService.getStats. This means that the task state will not
be inconsistent even if we don't clean it up.

Fixes: https://github.com/hazelcast/hazelcast/issues/11535