Fix CIES reordering WM and barrier & more (#542)
* Fix CIES to not reorder WM and barrier

* Change contract of Processor.finishSnapshotRestore: now it is called
each time the processor state is restored from snapshot. Previously, it
was only called if there was at least one key in the state.

* Simplifications in StreamEventJournalP: we don't need two places to
check for stale offset. We restore whatever offset there is in the
snapshot, it will fail later when trying to read from it.

* Initialize currInstream in ProcessorTasklet: having it uninitialized
caused the `tryProcess` method to by called once before
restoreFromSnapshot.

* Remove unused method from Traverser and outstanding TODOs

It could happen that even in exactly once mode, the WM and barriers
could be reordered with respect to each other.

This was caused by the fact that we could stop at a barrier in one queue,
that would cause the barrier to be emitted, but we continued to another
queue which drained a watermark, which would also cause a WM to be
emitted. After iterating all queues, we first checked for WM to emit and
then for barrier to emit, effectively causing the reordering. It was
fixed by moving the emit check inside to the loop and breaking the
iteration if we emitted either WM or barrier.