OperationParker.unpark simplification
In the loop a check is done if is the last parked operation from the queue,
and if it is, the key is removed from the parkQueueMap. But at the end of the loop,
you also know you have processed the last item.

So tis logic was pulled out of the loop to below the loop.